db_etl_pipeline/etl_rds_to_s3/requirements.txt
pylint
pytest
pyodbc
python-dotenv
pandas
boto3
awswrangler